# bajaj-admin/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medic 24x7 App",
        "description": "REST API's for Medic24x7 application",
        "termsOfService": "",
        "contact": {
            "email": "sayalib@leometric.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user/createEnquiry": {
            "post": {
                "tags": [
                    "Enquiry"
                ],
                "summary": "Insert Enquiry",
                "operationId": "enquiry",
                "parameters": [
                    {
                        "name": "partner_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/doctorlist": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "summary": "Doctor List",
                "operationId": "doctor",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/deleteAmbulance/{id}": {
            "delete": {
                "tags": [
                    "Ambulance"
                ],
                "summary": "Delete Ambulance record",
                "operationId": "ambulance",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/ambulancelist": {
            "get": {
                "tags": [
                    "Ambulance"
                ],
                "summary": "Ambulance List",
                "operationId": "ambulance",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/deleteBed/{id}": {
            "get": {
                "tags": [
                    "Beds"
                ],
                "summary": "Delete bed record",
                "operationId": "ambulance",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/deleteDoctor/{id}": {
            "delete": {
                "tags": [
                    "Doctor"
                ],
                "summary": "Delete Doctor record",
                "operationId": "ambulance",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}